chore(refactor): Remove duplicated files
