+++ src/routes/workouts/manage/overview/+page.svelte
@@ -5,7 +5,7 @@
 	import { toast } from 'svelte-sonner';
 	import LoaderCircle from 'virtual:icons/lucide/loader-circle';
 	import { workoutRunes } from '../workoutRunes.svelte';
-	import type { WorkoutExerciseInProgress } from '$lib/workoutFunctions';
+	import type { WorkoutExerciseInProgress } from '$lib/utils/workoutUtils';
 	import { goto, invalidate } from '$app/navigation';
 	import type { RouterInputs } from '$lib/trpc/router';
 	import * as Tabs from '$lib/components/ui/tabs';
@@ -34,7 +34,7 @@
 		const userBodyweight = workoutRunes.workoutData.userBodyweight;
 
 		const createData: RouterInputs['workouts']['create'] = {
-			workoutData: { ...workoutRunes.workoutData, userBodyweight  },
+			workoutData: { ...workoutRunes.workoutData, userBodyweight },
 			workoutExercises: workoutRunes.workoutExercises.map((ex, idx) => {
 				const { sets, ...exercise } = ex;
 				return { ...exercise, exerciseIndex: idx };
